Windows compatibility

+++ aintelope/training/lightning_trainer.py
@@ -3,6 +3,7 @@ import logging
 from pathlib import Path
 from collections import OrderedDict
 from datetime import timedelta
+import os
 
 from omegaconf import DictConfig
 import gym
@@ -103,7 +104,7 @@ class DQNLightning(LightningModule):
         states, actions, rewards, dones, next_states = batch
 
         state_action_values = (
-            self.net(states).gather(1, actions.unsqueeze(-1)).squeeze(-1)
+            self.net(states).gather(1, actions.unsqueeze(-1).long()).squeeze(-1)
         )
 
         with torch.no_grad():
